+++ R/check_spec.R
@@ -105,7 +105,7 @@ check_spec <- function(
       call = call
     )
   }
-  .check_spec_arg(spec, call = call)
+  spec <- .check_spec_arg(spec, call = call)
   .check_dataset_arg(spec, dataset, call = call)
 
   vars <- spec_variables(spec, dataset)