+++ R/decode_column.R
@@ -139,7 +139,7 @@ decode_column <- function(
       call = call
     )
   }
-  .check_spec_arg(spec, call = call)
+  spec <- .check_spec_arg(spec, call = call)
   .check_dataset_arg(spec, dataset, call = call)
   for (arg in c("from", "to")) {
     val <- get(arg)